experiments/cogvlm/requirements.txt
SwissArmyTransformer>=0.4.8
transformers>=4.33.1
xformers>=0.0.22
torch>1.10.0
torchvision
spacy>=3.6.0
pillow>=10.0.1
deepspeed>=0.11.0
seaborn
loguru~=0.7.2
accelerate
bitsandbytes
protobuf<=3.20.0